Move rviz worldloader to avoidance pkg
